third_party/blink/web_tests/virtual/indexeddb_sqlite/external/wpt/IndexedDB/idbindex_tombstones.any-expected.txt

This is a testharness.js-based test.
[FAIL] Forward iteration over an index in a readonly transaction
  assert_equals: expected 3 but got 6
[FAIL] Backward iteration over an index in a readonly transaction
  assert_equals: expected 3 but got 6
[FAIL] Forward iteration over an index in a readwrite transaction
  assert_equals: expected 3 but got 6
[FAIL] Backward iteration over an index in a readwrite transaction
  assert_equals: expected 3 but got 6
Harness: the test ran to completion.
